refactor: cleanup ProgressSliceWriter
